pyproject.toml
[project]
name = "trac-rpc"
description = "Modern, Pythonic and type-safe Trac RPC API client"
readme = "README.md"
authors = [
    { name = "Yury V. Zaytsev", email = "yury@shurup.com" }
]
repository = "https://github.com/zyv/trac-rpc-client"
keywords = ["trac", "rpc", "api", "pydantic", "httpx"]
license = "MIT"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dynamic = ["version"]
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "httpx",
]

[dependency-groups]
dev = [
    "hatch",
    "pytest-cov",
    "respx",
    "ruff",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/trac_rpc/__about__.py"

[tool.ruff]
line-length = 120
target-version = "py313"


[tool.ruff.lint]
select = [
    "C4", # flake8-comprehensions
    "COM", # flake8-commas
    "DTZ", # flake8-datetimez
    "E", # pycodestyle Error
    "F", # Pyflakes
    "I", # isort
    "LOG", # flake8-logging
    "PT", # flake8-pytest-style
    "PTH", # flake8-use-pathlib
    "Q", # flake8-quotes
    "PYI", # flake8-pyi
    "RSE", # flake8-raise
    "RUF", # ruff-specific rules
    "T10", # flake8-debugger
    "T20", # flake8-print
    "UP", # pyupgrade
    "W", # pycodestyle Warning
]

ignore = [
    "COM812", # comma consistency enforced by formatter
]

fixable = ["ALL"]
unfixable = []

[tool.pytest.ini_options]
log_cli = true
log_level = "DEBUG"

markers = [
    "integration: mark test as an integration test requiring a test Trac instance",
]

[tool.coverage.run]
branch = true
source = ["src"]
